requirements.txt
pandas>=1.5.0
tqdm>=4.64.0
selenium>=4.0.0
webdriver-manager>=3.8.0
beautifulsoup4>=4.11.0
requests>=2.28.0
google-generativeai>=0.3.0

# Dependências opcionais para dashboard web
# streamlit>=1.20.0
# plotly>=5.10.0
